Added keybind support for sorting current container. Moved some files around to have some more structure. Added translation file for english and put keybind information in there.
