+++ src/client/java/borknbeans/lightweightinventorysorting/LightweightInventorySortingClient.java
@@ -2,11 +2,29 @@ package borknbeans.lightweightinventorysorting;
 
 import borknbeans.lightweightinventorysorting.config.LightweightInventorySortingConfig;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
+import net.minecraft.client.option.KeyBinding;
+import net.minecraft.client.util.InputUtil;
+import org.lwjgl.glfw.GLFW;
 
 public class LightweightInventorySortingClient implements ClientModInitializer {
+
+	public static KeyBinding sortKeyBind;
+
 	@Override
 	public void onInitializeClient() {
 		// This entrypoint is suitable for setting up client-specific logic, such as rendering.
 		LightweightInventorySortingConfig.load();
+
+		registerKeyBindings();
+	}
+
+	private void registerKeyBindings() {
+		sortKeyBind = KeyBindingHelper.registerKeyBinding(new KeyBinding(
+				"key.lightweightinventorysorting.sort",
+				InputUtil.Type.KEYSYM,
+				GLFW.GLFW_KEY_R,
+				"category.lightweightinventorysorting.lightweightinventorysorting"
+		));
 	}
 }
