+++ src/client/java/borknbeans/lightweightinventorysorting/mixin/client/MixinGenericContainerScreen.java
@@ -1,6 +1,7 @@
 package borknbeans.lightweightinventorysorting.mixin.client;
 
 import borknbeans.lightweightinventorysorting.ContainerSortButton;
+import borknbeans.lightweightinventorysorting.LightweightInventorySortingClient;
 import borknbeans.lightweightinventorysorting.config.LightweightInventorySortingConfig;
 import net.minecraft.client.gui.DrawContext;
 import net.minecraft.client.gui.screen.ingame.GenericContainerScreen;
@@ -42,5 +43,14 @@ public abstract class MixinGenericContainerScreen extends HandledScreen<GenericC
         if (containerSortButton != null) {
             containerSortButton.render(context, mouseX, mouseY, delta);
         }
+    }
+
+    @Override
+    public boolean keyPressed(int keyCode, int scanCode, int modifiers) {
+        if (LightweightInventorySortingClient.sortKeyBind.matchesKey(keyCode, scanCode)) {
+            containerSortButton.onClick(0f, 0f); // Simulate a click
+        }
+
+        return super.keyPressed(keyCode, scanCode, modifiers);
     }
 }
