+++ src/client/java/borknbeans/lightweightinventorysorting/sorting/SortableSlotComparator.java
@@ -1,4 +1,4 @@
-package borknbeans.lightweightinventorysorting;
+package borknbeans.lightweightinventorysorting.sorting;
 
 import java.util.Comparator;
