+++ src/main/resources/assets/lightweight-inventory-sorting/lang/en_us.json
@@ -1,0 +1,4 @@
+{
+  "key.lightweightinventorysorting.sort": "Sort Current Container",
+  "category.lightweightinventorysorting.lightweightinventorysorting": "Lightweight Inventory Sorting"
+}
